Remove nodes that are not ready
This is the only way I can see to detect that a node is deleted.  It's
probably not perfect because nodes can be marked as not ready if they
get tainted or for any reason, really.  In addition, app pods may be
deployed on them while they're marked not ready.

But after an upgrade, we have to bounce the metrics endpoint to clear
the old nodes, which is a worse problem than what's described above.